Added logging and configuration enhancements
A new static method `log_new_line` was added to the `logger` class in `logger.h` for outputting a new line to the standard wide output stream. This method is now called in `sfbuild.cpp`'s `build_hub` method when `_DEBUG` is defined. A seemingly redundant line change was also made in `run_compiler` method of the same file.

In `sfconfigure.cpp`, handling for two new keys, `debugFlags` and `releaseFlags`, was added to the configuration parsing code. These keys, expected to be arrays of strings in the TOML configuration file, are appended to `m_configSetup.compilationList.debugFlags` and `m_configSetup.compilationList.releaseFlags` respectively, with a log message outputted for each new flag. A runtime error is thrown if either key's associated value is not an array of strings.